requirements/base.txt
Django==1.5.1
Fabric==1.6.0
South==0.7.6
django-bootstrap-toolkit==2.13.4
django-tinymce==1.5.1

